pyproject.toml
[tool.poetry]
name = "document-parser"
version = "0.1.0"
description = ""
authors = ["OchirLG <lidzhigoryaev.ochir47@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
html2text = "^2024.2.26"
pymupdf = "^1.25.3"
spire-doc = "^13.1.0"
ftfy = "^6.3.1"
pytest = "^8.3.5"
chardet = "^5.2.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
